refactor: extract worker and update notes
