+++ crates/scheduler/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod service;
 pub mod task;
+pub mod worker;
